change in message class

+++ server/maitre.go
@@ -43,7 +43,7 @@ func multicastSender() {
 			Genre: constantes.SYNC,
 			Id:    id,
 		}
-		tmaster := uint64(time.Now().UnixNano()) / uint64(time.Millisecond)
+		tmaster := time.Now().UnixNano() / int64(time.Millisecond)
 		message.SendMessage(sync.SimpleString(), conn)
 		follow_up := message.Message{
 			Genre: constantes.FOLLOW_UP,
